feat(sponsoredfarmlink): rename GetTokenLink to SponsoredFarmLink. Enable custom farm headings
