+++ src/components/earn/PoolCardTri/index.tsx
@@ -26,7 +26,7 @@ import {
   StyledActionsContainer,
   Button
 } from './PoolCardTri.styles'
-import GetTokenLink from '../FarmsPortfolio/GetTokenLink'
+import SponsoredFarmLink from '../../SponsoredFarmLink'
 import { StableSwapPoolName } from '../../../state/stableswap/constants'
 import { useSingleStableFarm } from '../../../state/stake/user-stable-farms'
 import PoolCardTriRewardText from './PoolCardTriRewardText'
@@ -116,7 +116,7 @@ const DefaultPoolCardtri = ({
 
       <AutoRow justifyContent="space-between">
         <PairContainer>
-          <GetTokenLink tokens={tokens} />
+          <SponsoredFarmLink tokens={tokens} farmID={version} />
           <MultipleCurrencyLogo currencies={currencies} size={20} />
           <ResponsiveCurrencyLabel currenciesQty={currenciesQty}>
             {friendlyFarmName ?? currencies.map(({ symbol }) => symbol).join('-')}
@@ -154,7 +154,6 @@ const DefaultPoolCardtri = ({
 type StablePoolCardTriProps = PoolCardTriProps & { stableSwapPoolName: StableSwapPoolName }
 
 const StableStakingPoolCardTRI = (props: StablePoolCardTriProps) => {
-
   const { version } = props
 
   const stakingInfo = useSingleStableFarm(Number(version), props.stableSwapPoolName)
